EventSampleRate is not integer but a floating point number

+++ src/base/signal_buffer.h
@@ -172,7 +172,7 @@ public:
 
     void removeEvent(uint32 event_id);
 
-    uint32 getEventSamplerate() const;
+    double getEventSamplerate() const;
